~ add TransformerWithoutSerializerTest to mark as legacy
