+++ tests/Entity/TransformerWithoutSerializerTest.php
@@ -1,0 +1,35 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\PageBundle\Tests\Entity;
+
+use Doctrine\Persistence\ManagerRegistry;
+use Sonata\PageBundle\Entity\Transformer;
+
+/**
+ * @group legacy
+ */
+final class TransformerWithoutSerializerTest extends TransformerTest
+{
+    protected function setUpTransformer(): void
+    {
+        $registry = $this->createMock(ManagerRegistry::class);
+
+        $this->transformer = new Transformer(
+            $this->snapshotManager,
+            $this->pageManager,
+            $this->blockManager,
+            $registry
+        );
+    }
+}
